# docs/swagger.yaml
definitions:
  category.Category:
    properties:
      _id:
        type: string
      base:
        $ref: '#/definitions/support.Base'
      code:
        type: string
      description:
        type: string
      name:
        type: string
      title:
        type: string
    type: object
  dashboard.ChartData:
    properties:
      all:
        $ref: '#/definitions/dashboard.Module'
      exclusive:
        $ref: '#/definitions/dashboard.Module'
      featured:
        $ref: '#/definitions/dashboard.Module'
      latest:
        $ref: '#/definitions/dashboard.Module'
      trending:
        $ref: '#/definitions/dashboard.Module'
    type: object
  dashboard.Dashboard:
    properties:
      all:
        $ref: '#/definitions/dashboard.Module'
      chartdata:
        $ref: '#/definitions/dashboard.ChartData'
      exclusive:
        $ref: '#/definitions/dashboard.Module'
      featured:
        $ref: '#/definitions/dashboard.Module'
      linechart:
        items:
          $ref: '#/definitions/news.Weekly'
        type: array
      news:
        items:
          $ref: '#/definitions/news.News'
        type: array
      sportcount:
        items:
          $ref: '#/definitions/news.SportCount'
        type: array
      trending:
        $ref: '#/definitions/dashboard.Module'
    type: object
  dashboard.Home:
    properties:
      exclusive:
        items:
          $ref: '#/definitions/news.News'
        type: array
      featured:
        items:
          $ref: '#/definitions/news.News'
        type: array
      latest:
        items:
          $ref: '#/definitions/news.News'
        type: array
      seo:
        $ref: '#/definitions/pages.Page'
      trending:
        items:
          $ref: '#/definitions/news.News'
        type: array
    type: object
  dashboard.Module:
    properties:
      name:
        type: string
      total:
        type: integer
    type: object
  news.Comment:
    properties:
      image:
        type: string
      message:
        type: string
      name:
        type: string
      time_ago:
        type: string
    type: object
  news.News:
    properties:
      _id:
        type: string
      base:
        $ref: '#/definitions/support.Base'
      caption:
        type: string
      code:
        type: string
      comments:
        items:
          $ref: '#/definitions/news.Comment'
        type: array
      content:
        type: string
      exclusive:
        type: boolean
      featured:
        type: boolean
      meta:
        type: string
      name:
        type: string
      picture:
        type: string
      sections:
        items:
          $ref: '#/definitions/news.Section'
        type: array
      sport:
        type: string
      time_ago:
        type: string
      title:
        type: string
      trending:
        type: boolean
      url:
        type: string
    type: object
  news.Section:
    properties:
      content:
        type: string
      highlight:
        type: boolean
      image:
        type: string
      name:
        type: string
    type: object
  news.SportCount:
    properties:
      count:
        type: integer
      sport:
        type: string
    type: object
  news.Weekly:
    properties:
      count:
        type: integer
      id:
        type: integer
    type: object
  pages.Page:
    properties:
      _id:
        type: string
      base:
        $ref: '#/definitions/support.Base'
      code:
        type: string
      content:
        type: string
      meta:
        type: string
      name:
        type: string
      title:
        type: string
    type: object
  support.Base:
    properties:
      created_At:
        type: string
      delete_At:
        type: string
      updated_At:
        type: string
    type: object
  support.HttpError:
    properties:
      code:
        type: integer
      errors:
        type: string
      message:
        type: string
      results:
        type: boolean
    type: object
...
info:
  contact: {}
paths:
...
  /api/categorys:
    delete:
      consumes:
      - application/json
      description: Delte item
      parameters:
      - description: code
        in: query
        name: code
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/category.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/support.HttpError'
      summary: Delte a category
      tags:
      - categorys
    get:
      consumes:
      - application/json
      description: Get item
      parameters:
      - description: code
        in: query
        name: code
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/category.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/support.HttpError'
      summary: Get a category
      tags:
      - categorys
    put:
      consumes:
      - application/json
      description: Update an new category item
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/category.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/support.HttpError'
      summary: Update an category
      tags:
      - categorys
  /api/dashboard:
    get:
      consumes:
      - application/json
      description: Get  Dashboard
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dashboard.Dashboard'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/support.HttpError'
      summary: Get  Dashboard
      tags:
      - Dashboard
...
  /home:
    get:
      consumes:
      - application/json
      description: Get  Home
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dashboard.Home'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/support.HttpError'
      summary: Get  Home
      tags:
      - Home
...
swagger: "2.0"